github-ci: add message in PR with coveralls link
Workflow 'debug_coverage' produces and uploads results to 'coveralls.io'
web site. Message about it can be shown in PR within each run was done.
This patch adds the ability to send message in available PR otherwise it
is skipped. Also added 'coverage.info' file to artifacts list.

Found that used 'coverallsapp/github-action' action checks if the
trigger was 'pull_request' used to run the workflow. And only in this
way it writes results message to PR. Previously 'pull_request' trigger
was blocked to avoid duplication with 'push' trigger. To make able to
run workflow by any single trigger and to send message to PR if it
exits, it was implemented the following logic:

  - run workflows on both triggers 'push' and 'pull_request';
  - for workflow with 'push' trigger check if PR exists then skip all
    jobs otherwise continue running;
  - for workflow with 'pull_request' trigger continue running;

Also found that due to opened issue coverallsapp/github-action#7 this
action is not currently ready to post results to github PR. Added
workaround, local print of message to PR with link to 'coveralls.io'
with results of the current branch.

To avoid of issue coverallsapp/github-action#55 [2], sources checkout
should be based on 2nd version and later, to support it newer docker
image is used tarantool/testing:debian-buster.

Closes #5644

[1]: https://github.com/coverallsapp/github-action/blob/master/src/run.ts#L38
[2]: https://github.com/coverallsapp/github-action/issues/55#issuecomment-644927165