test: fix flaky force_recovery test
In the test, there is a place, where it was checked that the amount
of valid data in snapshot in case when it was truncated is less than
in case we write garbage to it. Often it's really so, but depends on
the place of truncation/garbage location. Removed this check, because
on different systems, snapshot size is slightly different each time
you run test, so check will not pass every time.
Follow-up #5422