iproto: add check for user input of the number of iproto threads
At the moment, user can set any number of iproto threads, which leads
to incorrect behavior if the specified number of threads is less than
or equal to zero or too large. Add check for user input of the number
of iproto threads, value must be > 0 and less then or equal to 1000.

Follow-up #5645